pyproject.toml
[project]
name = "drone-swarms"
version = "0.1.0"
description = "JAX-accelerated multi-robot swarm simulation with MJX"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [
    {name = "JAVO"}
]

dependencies = [
    # JAX ecosystem - flexible version to work with existing installs
    # For GPU: pip install jax[cuda12]
    # For CPU: pip install jax (already installed in your env)
    "jax>=0.4.20",
    
    # MuJoCo + MJX (optional for now, can add later)
    # "mujoco>=3.2.0",
    # "mujoco-mjx>=3.2.0",
    
    # Neural networks & optimization
    "flax>=0.8.0",
    "optax>=0.2.0",
    
    # Graph Neural Networks
    "jraph>=0.0.6.dev0",
    
    # Utilities
    "chex>=0.1.0",
    "numpy>=1.24.0",
    "pyyaml>=6.0",
    "tqdm>=4.66.0",
    
    # Visualization
    "matplotlib>=3.7.0",
    
    # Experiment tracking (optional)
    # "wandb>=0.16.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["swarm"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
addopts = "-v --tb=short"

[tool.black]
line-length = 100
target-version = ["py310"]

[tool.ruff]
line-length = 100
select = ["E", "F", "I", "N", "W"]
ignore = ["E501"]
